just using class enums

+++ Projects/Moving PlayerBox/Moving PlayerBox/GameScene.h
@@ -25,9 +25,9 @@ namespace RB
 
 		void InitScene() override
 		{
-			objTree.CreateObj((int)GameObjType::background);
-			objTree.CreateObj((int)GameObjType::player, (int)ControllerType::PLAYER_CONTROLLER);
-			objTree.CreateObj((int)GameObjType::shitgroup, (int)ControllerType::SHIT_GENERATOR_CONTROLLER);
+			objTree.CreateObj(GameObjType::background);
+			objTree.CreateObj(GameObjType::player, ControllerType::PLAYER_CONTROLLER);
+			objTree.CreateObj(GameObjType::shitgroup, ControllerType::SHIT_GENERATOR_CONTROLLER);
 		}
 
 		void UpdateScene(olc::PixelGameEngine* ptrEngine, const GameData& gameData) override
@@ -37,10 +37,10 @@ namespace RB
 
 		void RenderScene(olc::PixelGameEngine* ptrEngine) override
 		{
-			objTree.GetObjType((int)GameObjType::background)->Render(ptrEngine, gameDecalLoader.GetDecal((int)GameSpriteType::background), RenderOffsetType::NONE);
-			objTree.GetObjType((int)GameObjType::player)->Render(ptrEngine, gameDecalLoader.GetDecal((int)GameSpriteType::playerbox_green), RenderOffsetType::BOTTOM_CENTER);
-			objTree.GetObjType((int)GameObjType::player)->Render(ptrEngine, gameDecalLoader.GetDecal((int)GameSpriteType::diamond_red), RenderOffsetType::CENTER_CENTER);
-			objTree.GetObjType((int)GameObjType::shitgroup)->Render(ptrEngine, gameDecalLoader.GetDecal((int)GameSpriteType::redbox_10x10), RenderOffsetType::CENTER_CENTER);
+			objTree.GetObjType(GameObjType::background)->Render(ptrEngine, gameDecalLoader.GetDecal((int)GameSpriteType::background), RenderOffsetType::NONE);
+			objTree.GetObjType(GameObjType::player)->Render(ptrEngine, gameDecalLoader.GetDecal((int)GameSpriteType::playerbox_green), RenderOffsetType::BOTTOM_CENTER);
+			objTree.GetObjType(GameObjType::player)->Render(ptrEngine, gameDecalLoader.GetDecal((int)GameSpriteType::diamond_red), RenderOffsetType::CENTER_CENTER);
+			objTree.GetObjType(GameObjType::shitgroup)->Render(ptrEngine, gameDecalLoader.GetDecal((int)GameSpriteType::redbox_10x10), RenderOffsetType::CENTER_CENTER);
 
 		}
 	};
